controller/python/requirements.txt
kopf>=1.37,<2
kubernetes>=29.0,<32
google-cloud-compute>=1.19,<2
google-cloud-network-connectivity>=2.5,<3
google-auth>=2.29,<3
